apply Amber's thediff, then separate None from 0
This makes sure that pos_weight and closed_class_weight are None when they
haven't been set and that no code checks for "if blah:", meaning that the else
branch will be taken *either* if blah is None *or* of it is 0. Instead, all
code checks for either "if blah is not None" or "if blah > 0", as appropriate.
It adds liberal assertions to make sure that exactly one of pos_weight and
closed_class_weight is non-None.